+++ src/main/java/racinggame/car/Car.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.car;
 
 import java.util.ArrayList;
 import java.util.List;
