+++ src/main/java/racinggame/car/Count.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.car;
 
 import racinggame.util.Validator;
